requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
cryptography==42.0.0
python-socketio==5.11.0
APScheduler==3.10.4
python-dotenv==1.0.0

